+++ app/src/main/java/com/money/manager/ex/common/CategoryListActivity.java
@@ -61,7 +61,7 @@ public class CategoryListActivity
             listFragment.mAction = intent.getAction();
             nestedListFragment.mAction = intent.getAction();
 
-            int requestId = intent.getIntExtra(KEY_REQUEST_ID, Constants.NOT_SET);
+            long requestId = intent.getLongExtra(KEY_REQUEST_ID, Constants.NOT_SET);
             listFragment.requestId = requestId;
             nestedListFragment.requestId = requestId;
         }
