+++ app/src/main/java/com/money/manager/ex/database/ISplitTransaction.java
@@ -28,12 +28,12 @@ import info.javaperformance.money.Money;
 public interface ISplitTransaction
     extends IEntity {
 
-    Integer getId();
-    void setId(int splitTransId);
+    Long getId();
+    void setId(long splitTransId);
     boolean hasId();
 
-    Integer getAccountId();
-    void setAccountId(int value);
+    Long getAccountId();
+    void setAccountId(long value);
 
     Money getAmount();
     void setAmount(Money splitTransAmount);
@@ -41,8 +41,8 @@ public interface ISplitTransaction
     String getNotes();
     void setNotes(String value);
 
-    Integer getCategoryId();
-    void setCategoryId(int categoryId);
+    Long getCategoryId();
+    void setCategoryId(long categoryId);
 
     TransactionTypes getTransactionType(TransactionTypes parentTransactionType);
     void setTransactionType(TransactionTypes value, TransactionTypes parentTransactionType);
