+++ app/src/main/java/com/money/manager/ex/domainmodel/SplitRecurringCategory.java
@@ -48,7 +48,7 @@ public class SplitRecurringCategory
     public static final String SPLITTRANSAMOUNT = "SPLITTRANSAMOUNT";
     public static final String NOTES = "NOTES";
 
-    public static SplitRecurringCategory create(int transactionId, int categoryId, int subcategoryId,
+    public static SplitRecurringCategory create(long transactionId, long categoryId, long subcategoryId,
                                                 TransactionTypes parentTransactionType, Money amount, String notes) {
         SplitRecurringCategory entity = new SplitRecurringCategory();
 
@@ -71,13 +71,13 @@ public class SplitRecurringCategory
     TransactionTypes transactionType;
 
     @Override
-    public Integer getId() {
-        return getInt(SPLITTRANSID);
+    public Long getId() {
+        return getLong(SPLITTRANSID);
     }
 
     @Override
-    public void setId(int value) {
-        setInt(SPLITTRANSID, value);
+    public void setId(long value) {
+        setLong(SPLITTRANSID, value);
     }
 
     @Override
@@ -86,18 +86,18 @@ public class SplitRecurringCategory
     }
 
     @Override
-    public Integer getAccountId() {
-        return getInt(ITransactionEntity.ACCOUNTID);
+    public Long getAccountId() {
+        return getLong(ITransactionEntity.ACCOUNTID);
     }
 
     @Override
-    public void setAccountId(int value) {
-        setInt(ITransactionEntity.ACCOUNTID, value);
+    public void setAccountId(long value) {
+        setLong(ITransactionEntity.ACCOUNTID, value);
     }
 
     @Override
-    public Integer getCategoryId() {
-        return getInt(CATEGID);
+    public Long getCategoryId() {
+        return getLong(CATEGID);
     }
 
     @Override
@@ -106,8 +106,8 @@ public class SplitRecurringCategory
     }
 
     @Override
-    public void setCategoryId(int categId) {
-        setInt(CATEGID, categId);
+    public void setCategoryId(long categId) {
+        setLong(CATEGID, categId);
     }
 
     @Override
@@ -131,12 +131,12 @@ public class SplitRecurringCategory
         DatabaseUtils.cursorDoubleToContentValuesIfPresent(c, contentValues, SPLITTRANSAMOUNT);
     }
 
-    public Integer getTransId() {
-        return getInt(TRANSID);
+    public Long getTransId() {
+        return getLong(TRANSID);
     }
 
-    public void setTransId(int value) {
-        setInt(TRANSID, value);
+    public void setTransId(long value) {
+        setLong(TRANSID, value);
     }
 
     public TransactionTypes getTransactionType(TransactionTypes parentTransactionType) {
