+++ app/src/main/java/com/money/manager/ex/investment/PortfolioCursorAdapter.java
@@ -46,7 +46,7 @@ public class PortfolioCursorAdapter
     }
 
     private final LayoutInflater mInflater;
-    private final HashMap<Integer, Integer> mHeadersAccountIndex;
+    private final HashMap<Long, Integer> mHeadersAccountIndex;
     private final SparseBooleanArray mCheckedPosition;
 
     @Override
@@ -72,7 +72,7 @@ public class PortfolioCursorAdapter
         PortfolioItemDataViewHolder holder = (PortfolioItemDataViewHolder) view.getTag();
 
         // header index
-        int accountId = cursor.getInt(cursor.getColumnIndex(StockFields.HELDAT));
+        long accountId = cursor.getInt(cursor.getColumnIndex(StockFields.HELDAT));
         if (!mHeadersAccountIndex.containsKey(accountId)) {
             mHeadersAccountIndex.put(accountId, cursor.getPosition());
         }
